[idl] Parse nullable types, types with parameter and array types

+++ compiler/test/data/idl/interfaceWithPrimitiveAttributes.d.kt
@@ -3,7 +3,7 @@ external abstract class interfaceName {
     var byteVar: Byte
     var shortVar: Short
     var intVar: Int
-    var longVar: Long
+    var longVar: Int
     var floatVar: Float
     var doubleVar: Double
 }
